+++ examples/SampleApp/App.tsx
@@ -1,10 +1,12 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { LogBox, Platform, useColorScheme } from 'react-native';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import { DarkTheme, DefaultTheme, NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { SafeAreaProvider, useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Chat, OverlayProvider, ThemeProvider, useOverlayContext } from 'stream-chat-react-native';
+import messaging from '@react-native-firebase/messaging';
+import notifee, { EventType } from '@notifee/react-native';
 import { AppContext } from './src/context/AppContext';
 import { AppOverlayProvider } from './src/context/AppOverlayProvider';
 import { UserSearchProvider } from './src/context/UserSearchContext';
@@ -35,10 +37,25 @@ import type {
   UserSelectorParamList,
 } from './src/types';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
+import { navigateToChannel, RootNavigationRef } from './src/utils/RootNavigation';
 
 LogBox.ignoreAllLogs(true);
 LogBox.ignoreLogs(['Non-serializable values were found in the navigation state']);
 console.assert = () => null;
+
+// when a channel id is set here, the intial route is the channel screen
+const initialChannelIdGlobalRef = { current: '' };
+
+notifee.onBackgroundEvent(async ({ detail, type }) => {
+  // user press on notification detected while app was on background on Android
+  if (type === EventType.PRESS) {
+    const channelId = detail.notification?.data?.channel_id;
+    if (channelId) {
+      navigateToChannel(channelId);
+    }
+    await Promise.resolve();
+  }
+});
 
 const Drawer = createDrawerNavigator();
 const Stack = createStackNavigator<StackNavigatorParamList>();
@@ -48,6 +65,51 @@ const App = () => {
   const colorScheme = useColorScheme();
   const streamChatTheme = useStreamChatTheme();
 
+  useEffect(() => {
+    const unsubscribeOnNotificationOpen = messaging().onNotificationOpenedApp((remoteMessage) => {
+      // Notification caused app to open from background state on iOS
+      const channelId = remoteMessage.data?.channel_id;
+      if (channelId) {
+        navigateToChannel(channelId);
+      }
+    });
+    // handle notification clicks on foreground
+    const unsubscribeForegroundEvent = notifee.onForegroundEvent(({ detail, type }) => {
+      if (type === EventType.PRESS) {
+        // user has pressed the foreground notification
+        const channelId = detail.notification?.data?.channel_id;
+        if (channelId) {
+          navigateToChannel(channelId);
+        }
+      }
+    });
+    notifee.getInitialNotification().then((initialNotification) => {
+      if (initialNotification) {
+        // Notification caused app to open from quit state on Android
+        const channelId = initialNotification.notification.data?.channel_id;
+        if (channelId) {
+          initialChannelIdGlobalRef.current = channelId;
+        }
+      }
+    });
+    messaging()
+      .getInitialNotification()
+      .then((remoteMessage) => {
+        if (remoteMessage) {
+          // Notification caused app to open from quit state on iOS
+          const channelId = remoteMessage.data?.channel_id;
+          if (channelId) {
+            // this will make the app to start with the channel screen with this channel id
+            initialChannelIdGlobalRef.current = channelId;
+          }
+        }
+      });
+    return () => {
+      unsubscribeOnNotificationOpen();
+      unsubscribeForegroundEvent();
+    };
+  }, []);
+
   return (
     <SafeAreaProvider
       style={{
@@ -55,6 +117,7 @@ const App = () => {
       }}
     >
       <NavigationContainer
+        ref={RootNavigationRef}
         theme={{
           colors: {
             ...(colorScheme === 'dark' ? DarkTheme : DefaultTheme).colors,
@@ -138,10 +201,17 @@ const HomeScreen = () => {
   const { overlay } = useOverlayContext();
 
   return (
-    <Stack.Navigator initialRouteName='ChatScreen'>
+    <Stack.Navigator
+      initialRouteName={initialChannelIdGlobalRef.current ? 'ChannelScreen' : 'ChatScreen'}
+    >
       <Stack.Screen component={ChatScreen} name='ChatScreen' options={{ headerShown: false }} />
       <Stack.Screen
         component={ChannelScreen}
+        initialParams={
+          initialChannelIdGlobalRef.current
+            ? { channelId: initialChannelIdGlobalRef.current }
+            : undefined
+        }
         name='ChannelScreen'
         options={{
           gestureEnabled: Platform.OS === 'ios' && overlay === 'none',