+++ package/src/components/Attachment/__tests__/Attachment.test.js
@@ -1,5 +1,8 @@
 import React from 'react';
 
+import { Linking } from 'react-native';
+
+import { act, renderHook } from '@testing-library/react-hooks';
 import { fireEvent, render, waitFor } from '@testing-library/react-native';
 import { v4 as uuidv4 } from 'uuid';
 
@@ -15,6 +18,7 @@ import {
 import { generateMessage } from '../../../mock-builders/generator/message';
 import { Attachment } from '../Attachment';
 import { AttachmentActions } from '../AttachmentActions';
+import { useGoToURL } from '../hooks/useGoToURL';
 
 const getAttachmentComponent = (props) => {
   const message = generateMessage();
@@ -118,6 +122,41 @@ describe('Attachment', () => {
     );
     await waitFor(() => {
       expect(getByTestId('card-attachment')).toBeTruthy();
+    });
+  });
+
+  it('should open the URL in the browser', async () => {
+    jest.spyOn(Linking, 'canOpenURL').mockImplementation(jest.fn().mockResolvedValue(true));
+
+    jest.spyOn(Linking, 'openURL').mockImplementation(jest.fn().mockResolvedValue(false));
+
+    const { result, waitFor } = renderHook(() => useGoToURL('www.google.com'));
+
+    const [error, openURL] = result.current;
+
+    await act(openURL);
+
+    await waitFor(() => {
+      expect(Linking.canOpenURL).toHaveBeenCalled();
+      expect(error).not.toBeTruthy();
+      expect(Linking.openURL).toHaveBeenCalled();
+    });
+  });
+
+  it('should open the URL in the browser when no url', async () => {
+    jest.spyOn(Linking, 'canOpenURL').mockImplementation(jest.fn().mockResolvedValue(false));
+
+    jest.spyOn(Linking, 'openURL').mockImplementation(jest.fn().mockResolvedValue(true));
+
+    const { result, waitForNextUpdate } = await renderHook(() => useGoToURL());
+    const [error, openURL] = result.current;
+
+    act(openURL);
+
+    await waitForNextUpdate(() => {
+      expect(Linking.canOpenURL).toHaveBeenCalled();
+      expect(error).toBeTruthy();
+      expect(Linking.openURL).not.toHaveBeenCalled();
     });
   });
 });