+++ package/src/contexts/themeContext/utils/theme.ts
@@ -80,6 +80,7 @@ export type MarkdownStyle = Partial<{
 export type Theme = {
   attachmentPicker: {
     bottomSheetContentContainer: ViewStyle;
+    durationText: TextStyle;
     errorButtonText: TextStyle;
     errorContainer: ViewStyle;
     errorText: TextStyle;
@@ -334,6 +335,7 @@ export type Theme = {
       unreadCountNotificationContainer: ViewStyle;
       unreadCountNotificationText: TextStyle;
       wrapper: ViewStyle;
+      chevronColor?: Color;
     };
     typingIndicatorContainer: ViewStyle;
   };
@@ -388,7 +390,7 @@ export type Theme = {
       eyeIcon: IconProps;
       /**
        * Available options for styling text:
-       * https://github.com/andangrd/react-native-markdown-package/blob/master/styles.js
+       * https://github.com/andangrd/react-native-markdown-package/blob/main/styles.js
        */
       markdown: MarkdownStyle;
       messageUser: TextStyle;
@@ -529,6 +531,7 @@ export type Theme = {
 export const defaultTheme: Theme = {
   attachmentPicker: {
     bottomSheetContentContainer: {},
+    durationText: {},
     errorButtonText: {},
     errorContainer: {},
     errorText: {},